refactor: element.dependantTypes resolver
