--- libs/backend/infra/adapter/neo4j/src/cypher/element/getElementTypes.cypher
@@ -1,4 +1,0 @@
-MATCH (this {id: $id})
--[:ARRAY_ITEM_TYPE|INTERFACE_FIELD|FIELD_TYPE|UNION_TYPE_CHILD*1..10]->(t)
-WHERE NOT t.id = $id AND NOT (t:PrimitiveType OR t:ReactNodeType OR t:ActionType OR t:RenderPropType)
-RETURN {id: t.id, __typename: LAST(labels(t))}
